README.md
The game is aim to hopefully teach players how to diagnose common PC problem. 
It’s design in a way that can be expand later on if needed. 
Please note that the demo is really barebone and is just a quick showcase. 

You can change the view using a dropdown menu. 
Selecting/Switching between component and remove/repair/putback selected component.
Instruction and debug lights led are there to help guide you.

You can download the demo in the  release section.